Depreciate CpcMeasureDataValidator and update PcfMeasureDataValidator
